float ->double

+++ ERP.API/Models/OrderProductCreateViewModel.cs
@@ -15,6 +15,6 @@ namespace ERP.API.Models
 
         public int? customer_order_id { get; set; }
         public int? op_discount { get; set; }
-        public float? op_total_value { get; set; }
+        public double? op_total_value { get; set; }
     }
 }
